// src/app/model/negocio.ts
export interface Negocio {
    id: number | null;
    id_usuario: number;
    nombre_empresa: string;
    categoria_empresa: string;
    telefono_empresa: string;
    direccion_empresa: string;
    horario_empresa: string;
    correo_empresa: string;
    descripcion_empresa: string;
    fotos_url_empresa: string;
    ine_url_representante: string;
}